requirements.txt
asgiref==3.9.1
beautifulsoup4==4.14.2
bs4==0.0.2
certifi==2025.8.3
charset-normalizer==3.4.3
Django==5.2.8
django-cleanup==9.0.0
django-multiselectfield==1.0.1
django-search-admin-autocomplete==0.2.1
django-simple-history==3.10.1
django-use-email-as-username==1.4.0
idna==3.10
numpy==2.3.5
OSGridConverter==0.1.3
pillow==11.3.0
requests==2.32.5
soupsieve==2.8
sqlparse==0.5.3
typing_extensions==4.15.0
Unidecode==1.4.0
urllib3==2.5.0
